+++ lizard_portal/templates/portals/esf-1.js
@@ -34,7 +34,7 @@
             title: 'Grafieken',
             flex: 1,
             xtype: 'multigraph',
-            open_map: function(workspace_slug, title, style, layer_name) {
+            open_map: function(workspace_slug, title, style, layer_name, extra) {
                 Ext.create('Ext.window.Window', {
                     modal: true,
                     title: title,
@@ -65,7 +65,6 @@
                         listeners: {
                             afterRender: function() {
                                 var me = this;
-                                debugger
                                 //me.callParent(arguments)
                                 me.setLoading(true);
                                 if (!this.init_workspace && this.autoLoadWorkspaceStore) {
@@ -95,7 +94,7 @@
                         }
                     },{
                         xtype: 'image',
-                        src: '/layers/wms/?FORMAT=image%2Fpng&TRANSPARENT=TRUE&STYLE='+ style + '&SERVICE=WMS&VERSION=1.1.1&REQUEST=GetLegendGraphic&SRS=EPSG%3A900913&BBOX=460467.38252035,6800747.0720688,640132.61747965,6894993.9279312&LAYER='+ layer_name +'&width=100&height=30',
+                        src: '/layers/wms/?FORMAT=image%2Fpng&TRANSPARENT=TRUE&STYLE='+ style + '&SERVICE=WMS&VERSION=1.1.1&REQUEST=GetLegendGraphic&SRS=EPSG%3A900913&BBOX=460467.38252035,6800747.0720688,640132.61747965,6894993.9279312&LAYER='+ layer_name +'&width=100&height=30' + extra,
                         height: 100
                     }]
                }).show()
@@ -105,17 +104,17 @@
             bbar: [{
                 text: 'PO4 op kaart',
                 handler: function (button) {
-                    button.up('panel').open_map('po4_map', 'PO4 op de kaart ', 'vss_ekr_value', 'vss:vss_track_records')
+                    button.up('panel').open_map('po4_map', 'PO4 op de kaart ', 'vss_track_record_PO4', 'vss:vss_track_records', '&CQL_FILTER=parameter_ident%3D\'PO4.bodem\'')
                 }
             },{
                 text: 'P op kaart',
                 handler: function (button) {
-                    button.up('panel').open_map('p_map', 'P op de kaart ', 'vss_ekr_value', 'vss:vss_track_records')
+                    button.up('panel').open_map('p_map', 'P op de kaart ', 'vss_track_record_Ptot', 'vss:vss_track_records', '&CQL_FILTER=parameter_ident%3D\'Ptot.bodem\'')
                 }
             },{
             text: 'AqMad op kaart',
             handler: function (button) {
-                button.up('panel').open_map('aqmad_map', 'AqMad op de kaart ', 'vss_ekr_value', 'vss:vss_area_value')
+                button.up('panel').open_map('aqmad_map', 'AqMad op de kaart ', 'vss_aqmad_Ptot', 'vss:vss_track_records', '&CQL_FILTER=parameter_ident%3D\'Ptot.z-score.water\'')
             }
         }],
             graphs: {% get_portal_template graphs-esf %}